CustomLauncher: Avoid crash when default qsb is reinflated
When onLayout was called by a different source (ex. Grid changes) it would crash the launcher because of the IllegalStateException thrown on the base context when goToState was called.

Instead of extending onLayout, to avoid unintentionally calling an unsupported context instance, we can call goToState only when told by QsbContainer

- Also add dummy SearchUiManager implementations so LauncherPreviewRenderer doesn't complain

Change-Id: I098c3bf545ca024eee251fb799f920ccf455d432